+++ crates/teepot/Cargo.toml
@@ -39,6 +39,8 @@ sha2.workspace = true
 signature.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
+tracing-log.workspace = true
+tracing-subscriber.workspace = true
 webpki-roots.workspace = true
 x509-cert.workspace = true
 zeroize.workspace = true
